Working on git issue #10 (https://github.com/SlotTeam/slot-cli/issues/10):
We have created the function isValidRootDir to know if the directory we are located is corresponds to a valid Slot project.

We are validating:

Not execute 'create command' over an existent slot project.
Execute 'create command' only if we are located on a valid slot project.
Initial testings were succeeded OK.. :smirk: , the issue will be closed soon after a few other pending tests.